refactor(Core): Move PasswordHashFactory to Mocks
This mock class is used by multiple tests leading to duplicate code
and therefor was moved to a separate Mock class.
