+++ src/Module/CoreModule/Tests/Infrastructure/UserInterface/Web/Form/Type/ChangePasswordTypeTest.php
@@ -13,11 +13,9 @@ namespace ParkManager\Module\CoreModule\Tests\Infrastructure\UserInterface\Web\F
 use Closure;
 use ParkManager\Module\CoreModule\Infrastructure\Security\ClientUser;
 use ParkManager\Module\CoreModule\Infrastructure\UserInterface\Web\Form\Type\Security\ChangePasswordType;
-use RuntimeException;
+use ParkManager\Module\CoreModule\Tests\Infrastructure\UserInterface\Web\Form\Type\Mocks\FakePasswordHashFactory;
 use Symfony\Component\Form\Test\Traits\ValidatorExtensionTrait;
 use Symfony\Component\Form\Test\TypeTestCase;
-use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
-use Symfony\Component\Security\Core\Encoder\PasswordEncoderInterface;
 
 /**
  * @internal
@@ -26,7 +24,7 @@ final class ChangePasswordTypeTest extends TypeTestCase
 {
     use ValidatorExtensionTrait;
 
-    /** @var EncoderFactoryInterface */
+    /** @var FakePasswordHashFactory */
     private $encoderFactory;
 
     protected function getExtensions(): array
@@ -38,35 +36,7 @@ final class ChangePasswordTypeTest extends TypeTestCase
 
     protected function setUp(): void
     {
-        $encoder = new class() implements PasswordEncoderInterface {
-            public function encodePassword($raw, $salt): string
-            {
-                return 'encoded(' . $raw . ')';
-            }
-
-            public function isPasswordValid($encoded, $raw, $salt): bool
-            {
-                return false;
-            }
-        };
-
-        $this->encoderFactory = new class($encoder) implements EncoderFactoryInterface {
-            private $encoder;
-
-            public function __construct($encoder)
-            {
-                $this->encoder = $encoder;
-            }
-
-            public function getEncoder($user): PasswordEncoderInterface
-            {
-                if ($user !== ClientUser::class) {
-                    throw new RuntimeException('Nope, that is not the right user.');
-                }
-
-                return $this->encoder;
-            }
-        };
+        $this->encoderFactory = new FakePasswordHashFactory();
 
         parent::setUp();
     }
