+++ src/Module/CoreModule/Tests/Infrastructure/UserInterface/Web/Form/Type/ConfirmPasswordResetTypeTest.php
@@ -15,13 +15,11 @@ use ParkManager\Module\CoreModule\Infrastructure\Security\ClientUser;
 use ParkManager\Module\CoreModule\Infrastructure\UserInterface\Web\Form\Type\Security\ConfirmPasswordResetType;
 use ParkManager\Module\CoreModule\Infrastructure\UserInterface\Web\Form\Type\Security\SecurityUserHashedPasswordType;
 use ParkManager\Module\CoreModule\Infrastructure\UserInterface\Web\Form\Type\Security\SplitTokenType;
+use ParkManager\Module\CoreModule\Tests\Infrastructure\UserInterface\Web\Form\Type\Mocks\FakePasswordHashFactory;
 use Rollerworks\Component\SplitToken\FakeSplitTokenFactory;
 use Rollerworks\Component\SplitToken\SplitToken;
-use RuntimeException;
 use Symfony\Component\Form\Test\Traits\ValidatorExtensionTrait;
 use Symfony\Component\Form\Test\TypeTestCase;
-use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
-use Symfony\Component\Security\Core\Encoder\PasswordEncoderInterface;
 use Symfony\Component\Translation\IdentityTranslator;
 
 /**
@@ -34,41 +32,13 @@ final class ConfirmPasswordResetTypeTest extends TypeTestCase
     /** @var FakeSplitTokenFactory */
     private $splitTokenFactory;
 
-    /** @var EncoderFactoryInterface */
+    /** @var FakePasswordHashFactory */
     private $encoderFactory;
 
     protected function setUp(): void
     {
-        $encoder = new class() implements PasswordEncoderInterface {
-            public function encodePassword($raw, $salt): string
-            {
-                return 'encoded(' . $raw . ')';
-            }
-
-            public function isPasswordValid($encoded, $raw, $salt): bool
-            {
-                return false;
-            }
-        };
-
         $this->splitTokenFactory = new FakeSplitTokenFactory();
-        $this->encoderFactory    = new class($encoder) implements EncoderFactoryInterface {
-            private $encoder;
-
-            public function __construct($encoder)
-            {
-                $this->encoder = $encoder;
-            }
-
-            public function getEncoder($user): PasswordEncoderInterface
-            {
-                if ($user !== ClientUser::class) {
-                    throw new RuntimeException('Nope, that is not the right user.');
-                }
-
-                return $this->encoder;
-            }
-        };
+        $this->encoderFactory    = new FakePasswordHashFactory();
 
         parent::setUp();
     }
